+++ metasvc-model/src/main/java/de/digitalcollections/model/exception/http/server/NotImplementedException.java
@@ -1,8 +1,10 @@
 package de.digitalcollections.model.exception.http.server;
 
+import org.zalando.problem.Problem;
+
 public class NotImplementedException extends HttpServerException {
 
-  public NotImplementedException(String methodKey, int status, String request) {
-    super(methodKey, status, request);
+  public NotImplementedException(String methodKey, int status, String request, Problem problem) {
+    super(methodKey, status, request, problem);
   }
 }